requirements.txt
Flask
requests
python-dotenv
flask_login
flask_sqlalchemy
flask_session
psycopg2-binary
Flask-SQLAlchemy